release 3.1.40 source code for net

+++ Services/Dcs/V2/Model/ListConfigTemplatesResponse.cs
@@ -20,7 +20,7 @@ namespace HuaweiCloud.SDK.Dcs.V2.Model
         /// 模板个数。
         /// </summary>
         [JsonProperty("template_num", NullValueHandling = NullValueHandling.Ignore)]
-        public decimal? TemplateNum { get; set; }
+        public int? TemplateNum { get; set; }
 
         /// <summary>
         /// 模板的详情数组。
